defringe cleanup (#96)
* Moved files around

* Changed imports so all defringe code is in defringe directory
