+++ stistools/defringe/fit1d.py
